% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_KRAS.R
\name{import_KRAS}
\alias{import_KRAS}
\title{Download and preprocess the raw dataset from Xue et al.}
\usage{
import_KRAS()
}
\value{
A \code{SingleCellExperiment} object
}
\description{
Download the dataset from GEO, filter, and create a
\code{SingleCellExperiment} object
}
\references{
Jenny Y. Xue, Yulei Zhao, Jordan Aronowitz, Trang T. Mai, Alberto Vides,
Besnik Qeriqi, Dongsung Kim, Chuanchuan Li, Elisa de Stanchina, Linas Mazutis,
 Davide Risso, and Piro Lito.
*Rapid non-uniform adaptation to conformation-specific KRAS(G12C) inhibition.*
Nature, 577(7790):421–425,jan  2020. ISSN  14764687. doi: 10.1038/s41586-019-1884-x
}
