% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_TGFB.R
\name{import_TGFB}
\alias{import_TGFB}
\title{Download and import the raw dataset from McFaline et al..}
\usage{
import_TGFB()
}
\description{
Download the dataset from GEO, filter, and create a
\code{SingleCellExperiment object}
}
\references{
Jos\'{e}  L.  McFaline-Figueroa,  Andrew  J.  Hill,  Xiaojie  Qiu,
Dana  Jackson,  Jay  Shendure,  and  Cole Trapnell.
*A pooled single-cell genetic screen identifies regulatory checkpoints in *
*the continuum of the epithelial-to-mesenchymal transition.*
Nature Genetics, 51(9):1389–1398, sep 2019.  ISSN 15461718.
doi:  10.1038/s41588-019-0489-5.
}
