% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kras}
\alias{kras}
\title{A \link[SingleCellExperiment]{SingleCellExperiment} object}
\format{
An object of class \code{SingleCellExperiment} with 100 rows and 10177 columns.
}
\source{
See the KRAS vignette to generate this data
}
\usage{
data(kras)
}
\description{
This dataset contains all the information used for the KRAS case studies
}
\references{
Jenny Y. Xue, Yulei Zhao, Jordan Aronowitz, Trang T. Mai, Alberto Vides,
Besnik Qeriqi, Dongsung Kim, Chuanchuan Li, Elisa de Stanchina, Linas Mazutis,
 Davide Risso, and Piro Lito.
*Rapid non-uniform adaptation to conformation-specific KRAS(G12C) inhibition.*
Nature, 577(7790):421–425,jan  2020. ISSN  14764687. doi: 10.1038/s41586-019-1884-x
}
\keyword{datasets}
