% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sling_reassign.R
\name{sling_reassign}
\alias{sling_reassign}
\title{Reassign cells after running slingshot with reassign = FALSE}
\usage{
sling_reassign(sds)
}
\arguments{
\item{sds}{Slingshot datasets}
}
\value{
the weights
}
\description{
Will return the weights after a final reassignment
}
